# ahms-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ayurvedic Hospital Management System — Phase 1 API (Auth, Roles, Departments, Patients, Doctors, Appointments, Dashboard).",
        "title": "AHMS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/encounters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "List encounters with filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by patient",
                        "name": "patient_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by department",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by doctor",
                        "name": "doctor_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Visit date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/encounters.EncounterResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reception creates a visit for an already-registered patient. Never creates a new patient.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Create an OPD encounter (visit) with an auto token",
                "parameters": [
                    {
                        "description": "Encounter",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/encounters.CreateEncounterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/encounters.EncounterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/encounters/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Get an encounter by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/encounters.EncounterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/encounters/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encounters"
                ],
                "summary": "Update an encounter's status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/encounters.UpdateEncounterStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/encounters.EncounterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "encounters.CreateEncounterRequest": {
            "type": "object",
            "required": [
                "department_id",
                "doctor_id",
                "patient_id"
            ],
            "properties": {
                "consultation_fee": {
                    "type": "number"
                },
                "department_id": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "encounter_type": {
                    "type": "string",
                    "enum": [
                        "OPD",
                        "IPD"
                    ]
                },
                "patient_id": {
                    "type": "string"
                },
                "referral_id": {
                    "type": "string"
                },
                "visit_date": {
                    "type": "string"
                },
                "visit_type": {
                    "type": "string",
                    "enum": [
                        "NEW",
                        "FOLLOW_UP"
                    ]
                }
            }
        },
        "encounters.EncounterConsultationItem": {
            "type": "object",
            "properties": {
                "ayurveda_fields": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "chief_complaints": {
                    "type": "string"
                },
                "clinical_notes": {
                    "type": "string"
                },
                "consultation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "diagnoses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/encounters.EncounterDiagnosisItem"
                    }
                },
                "diet_apathya": {
                    "type": "string"
                },
                "diet_pathya": {
                    "type": "string"
                },
                "examination": {
                    "type": "string"
                },
                "follow_up_date": {
                    "type": "string"
                },
                "history": {
                    "type": "string"
                },
                "treatment_plan": {
                    "type": "string"
                }
            }
        },
        "encounters.EncounterDiagnosisItem": {
            "type": "object",
            "properties": {
                "diagnosis": {
                    "type": "string"
                },
                "diagnosis_type": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "encounters.EncounterPrescriptionItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/encounters.EncounterPrescriptionItemLine"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "prescription_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "encounters.EncounterPrescriptionItemLine": {
            "type": "object",
            "properties": {
                "anupana": {
                    "type": "string"
                },
                "dispensed_qty": {
                    "type": "integer"
                },
                "dose": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "formulation": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "medicine": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "route": {
                    "type": "string"
                }
            }
        },
        "encounters.EncounterResponse": {
            "type": "object",
            "properties": {
                "consultation_fee": {
                    "type": "number"
                },
                "consultations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/encounters.EncounterConsultationItem"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "department_id": {
                    "type": "string"
                },
                "department_name": {
                    "type": "string"
                },
                "diagnoses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/encounters.EncounterDiagnosisItem"
                    }
                },
                "doctor_id": {
                    "type": "string"
                },
                "doctor_name": {
                    "type": "string"
                },
                "encounter_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "patient_name": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "prescriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/encounters.EncounterPrescriptionItem"
                    }
                },
                "referral_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token_number": {
                    "type": "integer"
                },
                "uhid": {
                    "type": "string"
                },
                "visit_date": {
                    "type": "string"
                },
                "visit_type": {
                    "type": "string"
                }
            }
        },
        "encounters.UpdateEncounterStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "REGISTERED",
                        "WAITING",
                        "IN_CONSULTATION",
                        "COMPLETED"
                    ]
                }
            }
        },
        "models.JSONB": {
            "type": "object",
            "additionalProperties": true
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}